fix: imports on move components to new folder
