+++ src/features/homePage/components/CreatePost.tsx
@@ -1,6 +1,5 @@
 import { type FC, type PropsWithChildren, useReducer, useState } from "react"
 import { Icon } from "~/components/Icon"
-import { ProfileAvatarImageUrl } from "~/components/profile/ProfileAvatarImageUrl"
 import { PrimaryButton } from "~/components/styledHTMLElements/StyledButtons"
 import { pollChoicesReducer, pollLengthReducer } from "../hooks"
 import { api } from "~/utils/api"
@@ -8,6 +7,7 @@ import toast from "react-hot-toast"
 import { CONFIG } from "~/config"
 import { type PostContent } from "../types"
 import { CreatePoll } from "./CreatePoll"
+import { ProfileAvatarImageUrl } from "~/features/profile"
 
 const INIT_POLL_LENGTH = {
 	days: 1,
